remove depricated svi/mcmc
