+++ docs/source/conf.py
@@ -27,9 +27,7 @@ os.environ['SPHINX_BUILD'] = '1'
 
 # HACK: This is to ensure that local functions are documented by sphinx.
 from numpyro.infer.mcmc import hmc  # noqa: E402
-from numpyro.infer.svi import svi  # noqa: E402
 hmc(None, None)
-svi(None, None, None, None)
 
 # -- Project information -----------------------------------------------------
